flash-kernel: Add support for Toshiba AC100
Add support for updating the kernel and initramfs in the Toshiba
AC100 boot partition located on the internal eMMC. This works by
introducing two new Boot-Type and Boot-Device-Minimal-Size fields.

The Boot-Type field determines what type of boot loader is used
on the machine. The intent of this is that we later get rid of
the duplicated machine lists in main(). The only value currently
understood is 'Android'.

The Boot-Device-Minimal-Size field specifies a minimal size the
boot partition must have. This is needed because the AC100 has
too such partitions, boot and recovery, and we want to use the
boot partition (the larger one).

Initially based on the initial code from Oliver Grawert
<ogra@ubuntu.com>, but completely rewritten to be vastly
more simple and match the changes in flash-kernel since
the original code.

Signed-off-by: Julian Andres Klode <jak@debian.org>