Call update-initramfs -u by default
This avoids maintenance of a long list of machine names in
flash-kernel-installer.postinst.

Almost all machines were listed in this case statement with identical
code at this point, except "Toshiba AC100 / Dynabook AZ" and
"Marvell DB-78x00-BP Development Board" because these are not (yet)
installable, but calling update-initramfs on these would certainly not
be incorrect.