Handle case when kernel == kfile check in boot_kernel_path handler.
This only applies to Method: generic.

Old behavior:

If the db entry specifies Boot-Kernel-Path, one of Dtb-Append, Machine-Id
or U-Boot-Kernel-Address must be specified which changes the value of $kernel.
When checking what to do with Boot-Kernel-Path, the script checks to see if
$kernel was changed from the original value ($kfile). If it did not change,
it does nothing. There is a TODO comment about handling symlinks, but this
does not make sense since this is the generic method, not the symlink method.

New behavior:

If $kernel was not changed, copy it to a temporary file before calling
backup_and_install. backup_and_install moves the temporary file, effectively
deleting it. Also applied same fix to initrd.

Use case:

On Raspberry Pi, the boot loader uses the vmlinux-* file directly, but it
needs to be renamed to kernel.img, which flash-kernel should be doing.

[ ijc -- fixed some typoes in commit log and added debian/changelog ]