Add RPI-Append for Raspberry Pi kernels.
The Raspberry Pi bootloader requires a "trailer" appended to the kernel image that
indicates if it supports device-tree or not. This is added by invoking the mkknlimg
command from <https://github.com/raspberrypi/tools/blob/master/mkimage/mkknlimg>.

Now, adding a line

    RPI-Append: yes

to a db file will cause mkknlimg to be called and the modified kernel image installed.