Disable the use of modprobe/udevadm in mtdblock() while running the testsuite.
The mtdblock() function calls modprobe and udevadm to make sure
that the relevant mtds are available.  This doesn't only happen
during "real" use, but also while running the testsuite during
the package build.  Potentially invasive actions such as loading
modules and calling into udev should not happen at build time,
and they are also not necessary for the testsuite as the
testsuite uses fake mtd data.  Therefore we disable the modprobe
and udevadm calls in mtdblock() when called from the testsuite.