Add support for Seagate NAS and Seagate Personal Cloud
Add support for Seagate NAS (2 and 4 bay) and Seagate Personal
Cloud (1 and 2 bay).

We load the kernel image to 0x2000000 and set the uImage load
address to 0x2000040 in order to make use of the U-Boot XIP
feature.  Thanks to Simon Guinot for this suggestion.

initrd_high needs to be set to 0xFFFFFFFF for u-boot to load the
ramdisk.