Fix script issue specific on a few linkstation devices
The issue occurs because a few linkstation devices use a script to detect
DTB shiped by kernel per version, and that script is also invoked after
kernel package upgrading or removing, which results in all kernel files
are removed including DTB. In this condition the linkstation specific script
fails to detect DTB, so flash-kernel hooks script fails, and finally triggers
kernel package upgrading or removing failing.

The fix is simply return a fake DTB as warning message instead of error,
when the kernel lib directory of the specific version doesn't exist.