Handle DTBs in vendor subdirectories (as found on arm64 kernels)
As of 3.19-rc1, DTBs on arm64 are placed in vendor subdirectories,
i.e. vendor/dtb.  Handle such DTBs by searching the whole kernel
hierarchy.

The DTB continues to be installed in dtbs/kvers/dtb (without a
vendor subdirectory) because the boot script looks for
for dtbs/${fk_kvers}/${fdtfile} and $fdtfile doesn't contain the
vendor subdir.

The current approach assumes that DTB filenames are unique when
disregarding the vendor dir.  I believe this is a pretty safe
assumption since DTB filenames start with the platform name.
If the assumption turns out to be incorrect in the future, we can
add the vendor dir to DTB-Id where needed and change the code to
look for vendor/dtb before looking for just dtb.

This approach was discussed here:
https://lists.debian.org/debian-boot/2016/07/msg00281.html