Only create DTB boot file on kernels that require DTB
Boot-DTB-Path can be specified to install the DTB to a file.  Some
machines, such as the SheevaPlug, contain both a DTB-Append-From and
a Boot-DTB-Path.

If DTB-Append-From is set, a verson check is performed to see if the
DTB is required and dtb_append is set to "yes" or "no" accordingly.
However, there is no version check for Boot-DTB-Path.  This can lead
to errors that the DTB couldn't be found on older kernels (i.e. older
than the version specified in DTB-Append-From).

Arguably, DTB-Append-From and Boot-DTB-Path together don't make sense
(why would you append the DTB _and_ install it to /boot), but it's
possible that users rely on this behaviour.

Therefore, only honour Boot-DTB-Path if $dtb_append is not "no".  If
it's "yes" or empty, we want to install the DTB to Boot-DTB-Path.
But if $dtb_append is "no", it means we're on an old kernel without
DTBs.

Debian bug #789886