flash-kernel: support relative paths for DTB-Id
Fixes: #860304

In U-Boot some boards expect the dtb file to be installed in a vendor
specific subdirectory.

With this patch relative paths specified in the DTB-Id are used for
this purpose.

The data base entry for the Hardkernel Odroid C2 is adjusted.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>

Minor changes to work with master: Vagrant Cascadian <vagrant@debian.org>